+++ docker/build_wei.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-UID=`id -u`
-GID=`id -g`
+export UID=`id -u`
+export GID=`id -g`
 
 SCRIPTPATH="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
