+++ requirements/requirements.txt
@@ -9,7 +9,7 @@ requests==2.31.0
 ulid-py==1.1.0
 pyzmq==25.1.1
 numpy==1.22.0
-opencv-python-headless==4.8.0.76
+opencv-python-headless
 redis==4.2.2
 
 pottery==3.0.0
