+++ wei/routers/__init__.py
@@ -1,0 +1,3 @@
+"""
+Contains all of the routes for the WEI server
+"""
